refactor: reuse table filter component in recurring actions feature
