+++ web/html/src/manager/systems/list-filter.tsx
@@ -1,8 +1,4 @@
-import { useState } from "react";
-
-import { Select } from "components/input";
-import { Form } from "components/input/Form";
-import { SelectSearchField } from "components/table/SelectSearchField";
+import { TableFilter } from "components/table/TableFilter";
 
 const SYSTEM_KIND_OPTIONS = [
   { value: "mgr_server", label: t("Manager Server") },
@@ -56,49 +52,6 @@ const allListOptions = [
   { value: "group_count", label: t("Groups") },
 ];
 
-const renderSearchField = (props) => {
-  const { field } = props;
-  const selectedOption = allListOptions.find((it) => it.value === field);
-  if (selectedOption?.filterOptions) {
-    return <SelectSearchField label={selectedOption.label} options={selectedOption.filterOptions} {...props} />;
-  }
-  return (
-    <div className="form-group">
-      <input
-        className="form-control"
-        value={props.criteria || ""}
-        placeholder={props.placeholder}
-        type="text"
-        onChange={(e) => props.onSearch?.(e.target.value)}
-        name={props.name}
-      />
-    </div>
-  );
-};
-
 export const SystemsListFilter = (props) => {
-  // Dummy model and onChange to reuse the Select component as it requires a Form
-  let model = {};
-  const onChange = () => {};
-
-  const [filterValue, setFilterValue] = useState(props.field || "");
-  const handleChangeSearchField = (value) => {
-    setFilterValue(value);
-    props.onSearchField?.(value);
-  };
-
-  return (
-    <Form model={model} onChange={onChange} title={t("Filter")} className="row">
-      <div className="col-sm-4">
-        <Select
-          name="filter"
-          placeholder={t("Select a filter")}
-          defaultValue={filterValue}
-          options={allListOptions}
-          onChange={(_name: string | undefined, value: string) => handleChangeSearchField(value)}
-        />
-      </div>
-      <div className="col-sm-6">{props.field && renderSearchField(props)}</div>
-    </Form>
-  );
+  return <TableFilter filterOptions={allListOptions} {...props} />;
 };
